+++ CHANGES.md
@@ -1,8 +1,9 @@
 # CHANGES
 
-## not-releaser-yet
+## not-released-yet
 
 - Pass some cli options via env variables
+- Add option to enable creation of intermediate directories
 
 ## 0.8.2